Add run-winfstest.ps1
We'll add another help that's supposed to run the tests. Unlinke
the existing cmd script, it's expecting python to be part of %PATH%
instead of attempting to locate it using registry keys that may not
be there.

At the same time, the Powershell arguments are much more easier to
read.